third_party/WebKit/LayoutTests/fast/dom/shadow/shadowradio.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div id="log"></div>
<span id="host"></span>
radios in normal dom [<input type="radio" name="r" id="a">]

<script>
var sr = document.getElementById('host').attachShadow({mode: 'open'});
sr.innerHTML = "radio inside shadow dom [<input type='radio' name='r' id='b'>][<input type='radio' name='r' id='c'>]";
var a = document.getElementById('a');
var b = sr.getElementById('b');
var c = sr.getElementById('c');

test(function() {
  a.checked = true;
  c.checked = true;
  assert_true(a.checked);
  assert_true(c.checked);
}, 'Radio buttons should not be grouped together across the shadow border.');

test(function() {
  b.checked = true;
  c.checked = true;
  assert_false(b.checked);
  assert_true(c.checked);
}, 'Radio buttons in a shadow tree should be grouped together.');

</script>
